pyproject.toml
[project]
name = "django-ninja-crudl"

[project.urls]
Homepage = "https://example.com"
Documentation = "https://readthedocs.org"
Repository = "https://github.com/me/spam.git"
Issues = "https://github.com/me/spam/issues"
Changelog = "https://github.com/me/spam/blob/master/CHANGELOG.md"

[tool.poetry]
name = "django-ninja-crudl"
version = "0.1.0"
description = "Converting Django models to Pydantic models"
authors = [
    "Jukka Hassinen <jukka.hassinen@gmail.com>"
]
license = "MIT"
readme = "README.md"
keywords = [
    "django",
    "pydantic",
    "models",
    "schema"
]
repository = "https://github.com/..."
homepage = "https://..."

classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Framework :: Flake8",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Quality Assurance",
  "Typing :: Typed",
]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "basedpyright>=1.21.0",
    "beartype>=0.19.0",
    "debugpy>=1.8.8",
    "django>=5.1.3",
    "django-ninja-extra>=0.21.4",
    "django-stubs>=5.1.1",
    "email-validator>=2.2.0",
    "hypothesis>=6.118.0",
    "import-linter>=2.1",
    "ipykernel>=6.29.5",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "ninja-schema>=0.13.6",
    "nox>=2024.10.9",
    "nplusone>=1.0.0",
    "pydantic>=2.9.2",
    "pylint-django>=2.6.1",
    "pyre-check>=0.9.23",
    "pyright>=1.1.388",
    "pytest>=8.3.3",
    "pytest-cov>=6.0.0",
    "pytest-django>=4.9.0",
    "pytest-instafail>=0.5.0",
    "pytest-testmon>=2.1.1",
    "pytest-watch>=4.2.0",
    "pytype>=2024.10.11",
    "rich>=13.9.4",
    "ruff>=0.7.3",
    "schemathesis>=3.38.5",
    "typeguard>=4.4.1",
    "wemake-python-styleguide>=0.19.2",
]
